# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RestReviews API - Restoranlar va Sharhlar",
        "description": "Mobile ilova uchun REST API. Restoranlarni kashf qilish, sharh qoldirish, sevimlilar va boshqalar.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gastronomic.webclub.uz",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://restreviews.test",
            "description": "Laragon Server"
        }
    ],
    "paths": {
        "/api/auth/send-code": {
            "post": {
                "tags": [
                    "🔐 Autentifikatsiya"
                ],
                "summary": "SMS tasdiqlash kodi yuborish",
                "description": "Telefon raqamiga SMS orqali 4 raqamli tasdiqlash kodini yuboradi. Kod 5 daqiqa davomida amal qiladi.",
                "operationId": "0f2ff5da3d5883ef973571776bfcbc27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Telefon raqami (998XXXXXXXXX formatida)",
                                        "type": "string",
                                        "example": "998901234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tasdiqlash kodi muvaffaqiyatli yuborildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tasdiqlash kodi yuborildi"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "998901234567"
                                        },
                                        "code": {
                                            "description": "Test rejimida qaytadi, production da null",
                                            "type": "string",
                                            "example": "1234",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatsiya xatosi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "phone": [
                                                    "The phone field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SMS yuborishda xatolik",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SMS yuborishda xatolik yuz berdi. Qaytadan urinib ko'ring."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/me": {
            "get": {
                "tags": [
                    "🔐 Autentifikatsiya"
                ],
                "summary": "Foydalanuvchi profilini olish",
                "description": "Autentifikatsiya qilingan foydalanuvchining profilini qaytaradi",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Profil ma'lumotlari",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "client": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Abdulla"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Valiyev"
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Abdulla Valiyev"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "998901234567"
                                                },
                                                "image_path": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya xatosi - token noto'g'ri yoki yo'q",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "🔐 Autentifikatsiya"
                ],
                "summary": "Tizimdan chiqish",
                "description": "Joriy tokenni bekor qiladi va foydalanuvchini tizimdan chiqaradi",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tizimdan chiqdingiz"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya xatosi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}